requirements.txt
attrs==25.3.0
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
filterpy==1.4.5
Flask==3.1.1
fonttools==4.58.5
idna==3.10
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
lxml==6.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
numpy==2.2.6
onvif_zeep==0.2.12
opencv-python==4.12.0.88
packaging==25.0
paho-mqtt==2.1.0
pillow==11.3.0
platformdirs==4.3.8
pyparsing==3.2.3
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.4
requests-file==2.1.0
requests-toolbelt==1.0.0
scipy==1.16.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.3
zeep==4.3.1
